+++ packages/orq-rc/RELEASES.md
@@ -9559,3 +9559,13 @@ Based on:
 - [typescript v4.10.0-rc.54] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.54] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.54 - packages/orq-rc
+
+## 2026-05-29 10:41:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.55] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.55] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.55 - packages/orq-rc